Commit en cours Mathilde

+++ src/main/java/dao/DaoReservation.java
@@ -2,6 +2,6 @@ package dao;
 
 import model.Reservation;
 
-public interface DaoReservation extends DaoGeneric<Reservation, Long> {
+public interface DaoReservation extends DaoGeneric<Reservation, Integer> {
 
 }
